+++ src/MethylPatternUniModelGainLoss.cpp
@@ -58,14 +58,14 @@ int MethylPatternUniModelGainLoss::learn(const ModelData& data, const vector<flo
             m_consensus_pattern.resize(s->pattern.size(), -1);
         }
 
-        for (unsigned int a=0; a<s->pattern.size(); a++) {
+        for (size_t a=0; a<s->pattern.size(); a++) {
             if (s->pattern[a] == 0)
                 n_zero[a] += factors[i];
             if (s->pattern[a] == 1)
                 n_one[a] += factors[i];
         }
     }
-    for (unsigned int i=0; i<n_zero.size(); i++) {
+    for (size_t i=0; i<n_zero.size(); i++) {
         if (n_zero[i] + n_one[i] > 0) {
             if (n_zero[i] >= n_one[i]) {
                 m_consensus_pattern[i] = 0;
@@ -80,7 +80,7 @@ int MethylPatternUniModelGainLoss::learn(const ModelData& data, const vector<flo
     float one_total_bits=0;
     float zero_error_bits=0;
     float zero_total_bits=0;
-    for (unsigned int index=0;index < m_consensus_pattern.size(); index++) {
+    for (size_t index=0;index < m_consensus_pattern.size(); index++) {
         if (m_consensus_pattern[index] == 0) {
             zero_error_bits += n_one[index];
             zero_total_bits += n_zero[index] + n_one[index];
@@ -144,7 +144,7 @@ int MethylPatternUniModelGainLoss::learn_gain_loss(const ModelData& data) {
             n_one.resize(s->pattern.size(),0);
         }
 
-        for (unsigned int a=0; a<s->pattern.size(); a++) {
+        for (size_t a=0; a<s->pattern.size(); a++) {
             if (s->pattern[a] == 0)
                 n_zero[a]++;
             if (s->pattern[a] == 1)
@@ -157,7 +157,7 @@ int MethylPatternUniModelGainLoss::learn_gain_loss(const ModelData& data) {
     float one_total_bits=0;
     float zero_error_bits=0;
     float zero_total_bits=0;
-    for (unsigned int index=0;index < m_consensus_pattern.size(); index++) {
+    for (size_t index=0;index < m_consensus_pattern.size(); index++) {
         if (m_consensus_pattern[index] == 0) {
             zero_error_bits += n_one[index];
             zero_total_bits+= n_zero[index] + n_one[index];
@@ -211,7 +211,7 @@ void MethylPatternUniModelGainLoss::simulate_sample(const Sample* original, Mode
     float r;
     const MethylPatternSample* os = (MethylPatternSample*)original;
     string s("");
-    for (unsigned int i=0; i<os->pattern.size(); i++) {
+    for (size_t i=0; i<os->pattern.size(); i++) {
         if (os->pattern[i] < 0)
             s += "*";
         else {
@@ -244,7 +244,7 @@ MethylPatternSample MethylPatternUniModelGainLoss::simulate(const MethylPatternS
     MethylPatternSample center(original, hamming_dist);
     for (int i = 0; i < sample_num; i++) {
         string s("");
-        for (unsigned int i = 0; i < center.pattern.size(); i++) {
+        for (size_t i = 0; i < center.pattern.size(); i++) {
             if (center.pattern[i] < 0)
                 s += "*";
             else {
@@ -275,7 +275,7 @@ float MethylPatternUniModelGainLoss::get_log_prob(const Sample* s) const {
     int correct_one=0;
     int errors_one=0;
     MethylPatternSample* ms = (MethylPatternSample*) s;
-    for (unsigned int i=0; i<ms->pattern.size(); i++) {
+    for (size_t i=0; i<ms->pattern.size(); i++) {
         if (0 == m_consensus_pattern[i]) {
             if (ms->pattern[i] == m_consensus_pattern[i]) {
                 correct_zero++;
